Refactoring to improve compile times
